mmusicc/ requirements.txt
mutagen==1.43.0
PyYAML==5.3
SQLAlchemy==1.3.12